Build 0.1

+++ src/SFA.DAS.AssessorService.Application.Api.UnitTests/WebAPI/Organisations/Queries/WhenSearchOrganisationSucceeds.cs
@@ -19,7 +19,7 @@ namespace SFA.DAS.AssessorService.Application.Api.UnitTests.WebAPI.Organisations
 
             _organisation = Builder<OrganisationResponse>.CreateNew().Build();
 
-            OrganisationQueryRepositoryMock.Setup(q => q.GetByUkPrn(Moq.It.IsAny<int>()))
+            OrganisationQueryRepositoryMock.Setup(q => q.GetByUkPrn(Moq.It.IsAny<long>()))
                 .Returns(Task.FromResult((_organisation)));
 
             _result = OrganisationQueryController.SearchOrganisation(10000000).Result;
